backend/pyproject.toml
[project]
name = "finance"
version = "0.1.0"
description = "Finance API backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.1",
    "pydantic[email]>=2.7.4",
    "pydantic-settings>=2.3.4",
    "motor>=3.4.0",
    "redis>=7.3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "bcrypt>=4.2.0",
    "pyjwt>=2.9.0",
    "alpaca-py>=0.43.2",
    "websockets>=13.0",
    "pytz>=2026.1.post1",
    "fastapi-limiter>=0.2.0",
]

[tool.pdm.scripts]
dev = "uvicorn finance.main:app --reload --port 8000"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
